app/requirements.txt
Django==3.0.7
django-cors-middleware==1.5.0
django-oauth-toolkit==1.3.2
djangorestframework==3.11.0
factory-boy==2.12.0
psycopg2-binary==2.8.5
Pygments==2.6.1
python-dotenv==0.13.0